rename plot.py to app.py. Add first test.
